+++ common-db/change-log.md
@@ -11,8 +11,7 @@
 10. 修复SqlUtils.convertNamedParamSql参数Map内的key存在而value为null时报错的问题
 11. SqlFile加入存储过程支持
 12. 修复getResultSetMap(ResultSet rs),getResultSetArray(ResultSet rs)参数越界的问题
-13. 加入ProcedureParameter,ProcedureOutParameter,ProcedureInOutParameter
-14. JdbcUtils加入存储过程的参数设置和存储过程的OUT Param的获取
+13. JdbcUtils加入存储过程的参数设置和存储过程的OUT Param的获取
 
 # 0.5.5 2022-8-11
 1. Dialect加入getKeywordLike(QueryPolicy)、getKeywordEq(QueryPolicy)、keywordsCase()方法